requirements.txt
discord.py
python-dotenv
gpt4all
